% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{INIT SERVICEBOT}
\usage{
init(servicebot_secret = NULL, secure_setup = FALSE,
  dev_mode = FALSE)
}
\arguments{
\item{servicebot_secret}{'Servicebot servicebot_secret'}

\item{secure_setup}{'HMAC string to be generated in embeds}

\item{dev_mode}{'Debug mode'}
}
\description{
INIT SERVICEBOT
}
