% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runPortalExample}
\alias{runPortalExample}
\title{Run examples of using shinyservicebot in a Shiny app}
\usage{
runPortalExample()
}
\description{
Run examples of using shinyservicebot in a Shiny app
}
\examples{
if (interactive()) {
  runPortalExample()
}
}
