% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runSignupExample}
\alias{runSignupExample}
\title{Run examples of using shinyservicebot in a Shiny app}
\usage{
runSignupExample()
}
\description{
Run examples of using shinyservicebot in a Shiny app
}
\examples{
if (interactive()) {
  runSignupExample()
}
}
