% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secure_setup}
\alias{secure_setup}
\title{Generate JWT}
\usage{
secure_setup(email, key)
}
\arguments{
\item{email}{Valid email adress}

\item{key}{Secret key}
}
\value{
JWT
}
\description{
Generate JWT
}
