% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyservicebot.R
\name{shinyservicebot}
\alias{shinyservicebot}
\title{Servicebot render embeds}
\usage{
shinyservicebot(servicebot_config = e$servicebot_config,
  dev_mode = e$dev_mode, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{servicebot_config}{This is the embed config for servicebot as described by servicebot.
See https://docs.servicebot.io/subscription-portal/subscription-management-portal/subscription-management-portal-configuration
for config options. See the \strong{examples} section below.}

\item{dev_mode}{For use in development. Set to FALSE in production}

\item{width}{With of html widget}

\item{height}{height of html widget}

\item{elementId}{elementId Use an explicit element ID for the widget (rather than an
automatically generated one). Ignored when used in a Shiny app.}
}
\description{
This function renders the html output for servicebot embeds
}
\examples{
# Init servicebot pr user login
shinyservicebot::init(dev_mode = TRUE)

# Config
servicebot_config = list (
 servicebot_id =  "E0OQN0P0Dort", # // LIVE & TEST MODE IDs ARE DIFFERENT
 service =   "Flat Subscription",
 email =  "shinyservicebot36io3ejbyf1pea7g61mxki-test@yopmail.com"
)

shinyservicebot::shinyservicebot(servicebot_config, width = "auto")

}
