% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyservicebot.R
\name{shinyservicebot_html}
\alias{shinyservicebot_html}
\title{Customize html}
\usage{
shinyservicebot_html(id, style, class, ...)
}
\arguments{
\item{id}{Id of div element}

\item{style}{Style element of html}

\item{class}{Class id of div element}

\item{...}{Other params}
}
\description{
Customize html
}
